Problem : App architecture inefficiencies Solution : move every ui to imports, currently currencyDetails page is moved
