+++ imports/ui/pages/currencyDetail/currencyDetail.html
