+++ imports/ui/pages/currencyDetail/features.html
