+++ imports/ui/pages/currencyDetail/redFlagCurrency.js
